requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.2.0
annotated-types==0.6.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncpg==0.29.0
attrs==23.2.0
Automat==22.10.0
babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.41.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clearml==1.14.4
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
crc32c==2.7
crispy-bootstrap5==2023.10
cryptography==42.0.7
cssselect==1.2.0
cycler==0.12.1
databases==0.9.0
databricks-sdk==0.20.0
datasets==2.19.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dj-database-url==2.1.0
djangify==1.0.0
Django==5.0.6
django-crispy-forms==2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
docker==7.0.0
donfig==0.8.1.post1
ecdsa==0.19.0
einops==0.7.0
environs==9.5.0
executing==2.0.1
fastapi==0.111.0
fastjsonschema==2.20.0
filelock==3.11.0
fire==0.5.0
Flask==3.0.3
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
furl==2.1.3
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.28.2
google-genai==0.5.0
graphene==3.2.2

graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.62.2
gunicorn==21.2
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httptools==0.6.0
httpx==0.27.0
huggingface-hub==0.23.0
hyperlink==21.0.0
idna==3.6
imagecodecs==2023.9.18
imageio==2.34.0
importlib_metadata==7.1.0
incremental==24.7.2
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.1.0

itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json-tricks==3.17.3
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyterlab==4.2.0
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lpips==0.1.4
lxml==5.1.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
medmnist==2.2.2
mistune==3.0.2
mlflow==2.11.3
mlflow-skinny==2.11.3
monai==1.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.0
ninja==1.11.1.1
nni==3.0
notebook==7.1.2
notebook_shim==0.2.4
numcodecs==0.13.1
onnx==1.15.0
openslide-python==1.3.1
optuna==3.5.0
packaging==23.2
pandas==2.2.2
parsel==1.8.1
parso==0.8.4
passlib==1.7.4
pathlib2==2.3.7.post1
peft==0.10.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.1
playwright==1.44.0
prettytable==3.10.0
prometheus_client==0.20.0
prompt_toolkit==3.0.43
propcache==0.3.1
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.2
pyamg==5.1.0
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.2
pydantic_core>=2.18.1
pydeck==0.8.1b0
pydicom==2.4.4
PyDispatcher==2.0.7
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.8.0
pynrrd==1.1.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1

python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.13
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
queuelib==1.6.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-file==2.1.0
responses==0.25.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
schema==0.7.7
scikit-image==0.22.0
scikit-learn==1.2.2

Scrapy==2.11.1
scrapy-playwright==0.0.32
seaborn==0.13.2
Send2Trash==1.8.3
service-identity==24.1.0
setuptools==67.8.0
SimpleITK==2.3.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
streamlit==1.32.2
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.4.0
tifffile==2023.9.26
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.1
triton==2.1.0
Twisted==24.3.0
typeguard==4.1.2
python-dateutil
types-python-dateutil
typing-inspection==0.4.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
w3lib==2.1.2
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.2
wheel==0.43.0
whitenoise==6.6.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xgboost==2.0.3
xxhash==3.4.1
yarl==1.9.4
zarr==2.16.1
zipp==3.18.1
zope.interface==6.3
